+++ wheelfile.py
@@ -1611,6 +1611,7 @@ class WheelFile:
               filename: Union[str, Path],
               arcname: Optional[str] = None,
               compress_type: Optional[int] = None,
+              compresslevel: Optional[int] = None,
               *, recursive: bool = True, resolve: bool = True,
               skipdir: bool = True) -> None:
         """Add the file to the wheel.
@@ -1629,6 +1630,10 @@ class WheelFile:
 
         compress_type
             Same as in `zipfile.ZipFile.write`. Overrides the `compression`
+            parameter given to `__init__`.
+
+        compresslevel
+            Same as in `zipfile.ZipFile.write`. Overrides the `compresslevel`
             parameter given to `__init__`.
 
         recursive
@@ -1657,7 +1662,8 @@ class WheelFile:
         """
         if resolve and arcname is None:
             arcname = resolved(filename)
-        self._write_to_zip(filename, arcname, skipdir, compress_type)
+        self._write_to_zip(filename, arcname, skipdir, compress_type,
+                           compresslevel)
 
         if recursive:
             common_root = str(filename)
@@ -1674,9 +1680,10 @@ class WheelFile:
                         common_root, root, name, root_arcname
                     )
                     self._write_to_zip(filepath, arcpath, skipdir,
-                                       compress_type)
+                                       compress_type, compresslevel)
 
-    def _write_to_zip(self, filename, arcname, skipdir, compress_type):
+    def _write_to_zip(self, filename, arcname, skipdir, compress_type,
+                      compresslevel):
         zinfo = zipfile.ZipInfo.from_file(
             filename, arcname, strict_timestamps=self._strict_timestamps
         )
@@ -1688,7 +1695,7 @@ class WheelFile:
         else:
             with open(filename, 'br') as f:
                 data = f.read()
-        self._zip.writestr(zinfo, data, compress_type)
+        self._zip.writestr(zinfo, data, compress_type, compresslevel)
         self.refresh_record(zinfo.filename)
 
     @staticmethod
@@ -1711,7 +1718,8 @@ class WheelFile:
     def writestr(self,
                  zinfo_or_arcname: Union[zipfile.ZipInfo, str],
                  data: Union[bytes, str],
-                 compress_type: Optional[int] = None) -> None:
+                 compress_type: Optional[int] = None,
+                 compresslevel: Optional[int] = None) -> None:
         """Write given data into the wheel under the given path.
 
         Updates the wheel record, if the record is being kept.
@@ -1731,6 +1739,12 @@ class WheelFile:
             parameter given to `__init__`. If the first parameter is a
             `ZipInfo` object, the value its `compress_type` field is also
             overriden.
+
+        compresslevel
+            Same as in `zipfile.ZipFile.writestr`. Overrides the `compresslevel`
+            parameter given to `__init__`. If the first parameter is a
+            `ZipInfo` object, the value its `compresslevel` field is also
+            overriden.
         """
 
         # XXX: ZipFile.writestr() does not normalize arcpaths the same way
@@ -1742,7 +1756,7 @@ class WheelFile:
             else zinfo_or_arcname
         )
 
-        self._zip.writestr(zinfo_or_arcname, data, compress_type)
+        self._zip.writestr(zinfo_or_arcname, data, compress_type, compresslevel)
         self.refresh_record(arcname)
 
     # TODO: compression args?
@@ -1752,6 +1766,7 @@ class WheelFile:
     def write_data(self, filename: Union[str, Path],
                    section: str, arcname: Optional[str] = None,
                    compress_type: Optional[int] = None,
+                   compresslevel: Optional[int] = None,
                    *, recursive: bool = True, resolve: bool = True,
                    skipdir: bool = True) -> None:
         """Write a file to the .data directory under a specified section.
@@ -1779,6 +1794,10 @@ class WheelFile:
 
         compress_type
             Same as in `zipfile.ZipFile.write`. Overrides the `compression`
+            parameter given to `__init__`.
+
+        compresslevel
+            Same as in `zipfile.ZipFile.write`. Overrides the `compresslevel`
             parameter given to `__init__`.
 
         recursive
@@ -1815,8 +1834,8 @@ class WheelFile:
         arcname = self._distinfo_path(section + '/' + arcname.lstrip('/'),
                                       kind='data')
 
-        self.write(filename, arcname, compress_type, recursive=recursive,
-                   resolve=resolve, skipdir=skipdir)
+        self.write(filename, arcname, compress_type, compresslevel,
+                   recursive=recursive, resolve=resolve, skipdir=skipdir)
 
     # TODO: compression args?
     # TODO: drive letter should be stripped from the arcname the same way
@@ -1825,7 +1844,8 @@ class WheelFile:
     def writestr_data(self, section: str,
                       zinfo_or_arcname: Union[zipfile.ZipInfo, str],
                       data: Union[bytes, str],
-                      compress_type: Optional[int] = None) -> None:
+                      compress_type: Optional[int] = None,
+                      compresslevel: Optional[int] = None) -> None:
         """Write given data to the .data directory under a specified section.
 
         This method is a handy shortcut for writing into
@@ -1855,6 +1875,12 @@ class WheelFile:
             parameter given to `__init__`. If the first parameter is a
             `ZipInfo` object, the value its `compress_type` field is also
             overriden.
+
+        compresslevel
+            Same as in `zipfile.ZipFile.writestr`. Overrides the `compresslevel`
+            parameter given to `__init__`. If the first parameter is a
+            `ZipInfo` object, the value its `compresslevel` field is also
+            overriden.
         """
         self._check_section(section)
 
@@ -1867,12 +1893,13 @@ class WheelFile:
         arcname = self._distinfo_path(section + '/' + arcname.lstrip('/'),
                                       kind='data')
 
-        self.writestr(arcname, data, compress_type)
+        self.writestr(arcname, data, compress_type, compresslevel)
 
     # TODO: Lazy mode should permit writing meta here
     def write_distinfo(self, filename: Union[str, Path],
                        arcname: Optional[str] = None,
                        compress_type: Optional[int] = None,
+                       compresslevel: Optional[int] = None,
                        *, recursive: bool = True, resolve: bool = True,
                        skipdir: bool = True) -> None:
         """Write a file to `.dist-info` directory in the wheel.
@@ -1897,6 +1924,10 @@ class WheelFile:
 
         compress_type
             Same as in `zipfile.ZipFile.write`. Overrides the `compression`
+            parameter given to `__init__`.
+
+        compresslevel
+            Same as in `zipfile.ZipFile.write`. Overrides the `compresslevel`
             parameter given to `__init__`.
 
         recursive
@@ -1947,13 +1978,14 @@ class WheelFile:
 
         arcname = self._distinfo_path(arcname)
 
-        self.write(filename, arcname, compress_type, recursive=recursive,
-                   skipdir=skipdir)
+        self.write(filename, arcname, compress_type, compresslevel,
+                   recursive=recursive, skipdir=skipdir)
 
     # TODO: Make sure fields of given ZipInfo objects are propagated
     def writestr_distinfo(self, zinfo_or_arcname: Union[zipfile.ZipInfo, str],
                           data: Union[bytes, str],
-                          compress_type: Optional[int] = None) -> None:
+                          compress_type: Optional[int] = None,
+                          compresslevel: Optional[int] = None) -> None:
         """Write given data to the .dist-info directory.
 
         This method is a handy shortcut for writing into
@@ -1982,6 +2014,12 @@ class WheelFile:
             `ZipInfo` object, the value its `compress_type` field is also
             overriden.
 
+        compresslevel
+            Same as in `zipfile.ZipFile.writestr`. Overrides the `compresslevel`
+            parameter given to `__init__`. If the first parameter is a
+            `ZipInfo` object, the value its `compresslevel` field is also
+            overriden.
+
         Raises
         ------
         ProhibitedWriteError
@@ -2001,7 +2039,7 @@ class WheelFile:
             )
 
         arcname = self._distinfo_path(arcname.lstrip('/'))
-        self.writestr(arcname, data, compress_type)
+        self.writestr(arcname, data, compress_type, compresslevel)
 
     @staticmethod
     def _check_section(section):